Added Request Sent page / linked through forgot password submit button
Added "BackToLoginButton" component (BackToLogin.js). The component is reused and this makes it easier to keep them the same

Request Sent page includes text notification that the user should receive an email, a BackToLogin button and a "Resend Link" button. A section for an image is also included but no image has yet been added